+++ src/main/java/org/broadinstitute/sequel/infrastructure/bsp/plating/SeqWorkRequestAliquot.java
@@ -5,22 +5,21 @@ package org.broadinstitute.sequel.infrastructure.bsp.plating;
  * Interface for mapping between Squid Sequencing Work Requests and BSP Plating
  * Work Requests. This describes the way GSP PMs would like aliquots made from
  * BSP stock samples.
- * 
+ *
  * @author mcovarr
- * 
  */
 public class SeqWorkRequestAliquot implements Plateable {
-    
+
     private String bspStockSampleId;
-    
-    private Double desiredDNAConcentration;
-    
-    private Double desiredVolume;
-        
+
+    private Float desiredDNAConcentration;
+
+    private Float desiredVolume;
+
     private String quoteId;
-    
-    
-    public SeqWorkRequestAliquot(String bspStockSampleId, Double desiredVolume, Double desiredDNAConcentration, String quoteId) {
+
+
+    public SeqWorkRequestAliquot(String bspStockSampleId, Float desiredVolume, Float desiredDNAConcentration, String quoteId) {
         setBspStockSampleId(bspStockSampleId);
         setDesiredVolume(desiredVolume);
         setDesiredDNAConcentration(desiredDNAConcentration);
@@ -28,69 +27,65 @@ public class SeqWorkRequestAliquot implements Plateable {
     }
 
     /**
-     * 
      * @return The Sample ID (e.g. SM-12CO4) of the BSP stock sample from
-     * which the aliquot should be made. 
+     *         which the aliquot should be made.
      */
     public String getBspStockSampleId() {
         return bspStockSampleId;
     }
-    
+
     /**
-     * 
      * @return Stringified quote to use for this sample/aliquot
      */
     String getQuoteId() {
         return quoteId;
     }
-    
+
     /**
-     * 
      * @return desired aliquot concentration (ng/uL)
      */
-    public Double getDesiredDNAConcentration() {
+    public Float getDesiredDNAConcentration() {
         return desiredDNAConcentration;
     }
 
     /**
-     * 
      * @return desired aliquot volume (uL)
-     */    
-    public Double getDesiredVolume() {
+     */
+    public Float getDesiredVolume() {
         return desiredVolume;
     }
-    
-    
+
+
     public void setBspStockSampleId(String bspStockSampleId) {
-        
+
         if (bspStockSampleId == null)
             throw new RuntimeException("BSP stock sample ID can not be null");
-        
+
         this.bspStockSampleId = bspStockSampleId;
     }
 
-    public void setDesiredDNAConcentration(Double desiredDNAConcentration) {
-        
+    public void setDesiredDNAConcentration(Float desiredDNAConcentration) {
+
         if (desiredDNAConcentration == null)
             throw new RuntimeException("Desired DNA concentration can not be null");
-        if (desiredDNAConcentration <= 0) 
+        if (desiredDNAConcentration <= 0)
             throw new RuntimeException("Desired DNA concentration must be positive");
-        
+
         this.desiredDNAConcentration = desiredDNAConcentration;
     }
 
-    public void setDesiredVolume(Double desiredVolume) {
+    public void setDesiredVolume(Float desiredVolume) {
         if (desiredVolume == null)
             throw new RuntimeException("Desired volume can not be null");
         if (desiredVolume < 1)
             throw new RuntimeException("Desired volume must be >= 1.0 uL");
-        
+
         this.desiredVolume = desiredVolume;
     }
 
 
     public void setQuoteId(String quoteId) {
-        if (quoteId == null) 
+        if (quoteId == null)
             throw new RuntimeException("Plating quote ID can not be null");
         this.quoteId = quoteId;
     }
@@ -113,12 +108,12 @@ public class SeqWorkRequestAliquot implements Plateable {
     }
 
     @Override
-    public Double getVolume() {
+    public Float getVolume() {
         return getDesiredVolume();
     }
 
     @Override
-    public Double getConcentration() {
+    public Float getConcentration() {
         return getDesiredDNAConcentration();
     }
 }
